requirements.txt
# Core ML
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Explainability
shap>=0.44.0

# Experiment tracking
mlflow>=2.9.0

# Data drift monitoring
evidently>=0.4.0

# App
streamlit>=1.28.0
plotly>=5.18.0

# Config
PyYAML>=6.0

# Testing & code quality
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
joblib>=1.3.0